Création, liste, details et suppresion OK (manque validation edit)

+++ CarShare/CarShare/BO/Emplacement.cs
@@ -11,11 +11,10 @@ namespace CarShare.BO
     {
         public string Intitule { get; set; }
         public string Description { get; set; }
-        public long Latitude { get; set; }
-        public long Longitude { get; set; }
-        public virtual Ecole Ecole { get; set; }
+        public float Latitude { get; set; }
+        public float Longitude { get; set; }
+        public virtual Ecole Ecole{ get; set; }
         public virtual Arret Arret { get; set; }
-
         private int id;
         [Key]
         public int Id
